Introduce namespaced symbols
This is a first cut at implementing #277. It seems to work
(in particular, it solves the problem discussed in #273),
but it has a few rough edges.

First, the constructor for namespaced symbols in hy.models.symbol is a
bit hackish. I don't really know why the straightforward solution (add
an optional argument to HySymbol's __init__ and set the attribute
there) doesn't work. This requires some exploration.

Second, all symbols in a quasiquote currently get a namespace
attached, even those referring to built-in special forms.
That works because the namespace part is ignored everywhere
except where the symbols ends up being evaluated, but it's
not what I call elegant.

Third, the modules referred to by a resolving a namespaced
symbol are imported under their plain names, which can cause
conflicts if the same symbol is used for some other purpose.
However, this problem occurs with all the auto-imports the
compiler does, so it should be fixed elsewhere.